requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
coremltools>=6.3
distro==1.9.0
dotenv==0.9.9
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.3.2
markdown==3.8
mlx>=0.0.5
mlx-core>=0.0.3
numpy==2.2.5
openai==1.75.0
pandas==2.2.3
pillow==11.2.1
platform-detect>=0.1.0
pydantic==2.11.3
pydantic-core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-frontmatter==1.1.0
pytrends==4.9.2
pytz==2025.2
pyyaml==5.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
